PRD-2948: Updated the launcher to be able to perform system property mappings from environment variables. This makes sure we do not need to employ Shell-Scripts for this task.
Example:
system-property.org.pentaho.home=${HOME}

Defines a system property "org.pentaho.home" and sets it to the value of the Environment variable HOME.

In addition to that: Added a option so that the launcher can disable an previously installed security manager. Security managers make report processing about 50% slower than normal even if a "allow-all" policy is in place. This must be explicitly enabled in the launcher.properties file.

Example:
uninstall-security-manager=true


git-svn-id: svn://source.pentaho.org/svnroot/pentaho-commons/pentaho-application-launcher/trunk@43888 5d404267-db13-0410-8476-b1f6b398e63a